Directly view existing explanation (#57)
When clicking on the "Explain Error" button when viewing the console
output and an explanation already exists a intermediate pseudo dialog
was shown to either show the explanation, regenerate or cancel. I think
that is not a good UI experience.
An existing explanation is now directly shown when clicking in "Explain
Error". The card now has 2 buttons in the top right that allow to
regenerate and to close the explanation.
Even when the explainer is disabled after an explanation has been
created for a run, the button will be available. But re-explaining will
not be available in that case.
The change also prevents that the footer injects anything to a page or
loads the js when not on a console url.

fixes #52 
<!-- Please describe your pull request here. -->


After:
<img width="1052" height="233" alt="image"
src="https://github.com/user-attachments/assets/59c99383-b1c4-4a9f-ae62-9828122db5c6"
/>

### Testing done
Interactive testing and added unit tests that ensure the div is properly
injected.


<!-- Comment:
Provide a clear description of how this change was tested.
At minimum this should include proof that a computer has executed the
changed lines.
Ideally this should include an automated test or an explanation as to
why this change has no tests.
Note that automated test coverage is less than complete, so a successful
PR build does not necessarily imply that a computer has executed the
changed lines.
If automated test coverage does not exist for the lines you are
changing, you must describe the scenario(s) in which you manually tested
the change.
For frontend changes, include screenshots of the relevant page(s) before
and after the change.
For refactoring and code cleanup changes, exercise the code before and
after the change and verify the behavior remains the same.
-->

### Submitter checklist
- [x] Make sure you are opening from a **topic/feature/bugfix branch**
(right side) and not your main branch!
- [x] Ensure that the pull request title represents the desired
changelog entry
- [x] Please describe what you did
- [x] Link to relevant issues in GitHub or Jira
- [ ] Link to relevant pull requests, esp. upstream and downstream
changes
- [ ] Ensure you have provided tests that demonstrate the feature works
or the issue is fixed

<!--
Put an `x` into the [ ] to show you have filled the information.
The template comes from
https://github.com/jenkinsci/.github/blob/master/.github/pull_request_template.md
You can override it by creating .github/pull_request_template.md in your
own repository
-->